machine-learning/requirements.txt
# Core Requirements
torch==2.1.2
torchvision==0.16.2
pillow==10.1.0
flask==3.0.2
numpy==1.26.4
